pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cbio-nav-agent"
version = "0.1.0"
description = "Claude agent that connects to the cBioPortal MCP server and exposes an OpenAI-compatible API."
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.28.0",
    "fastapi>=0.111.0",
    "mcp>=0.1.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.36.0",
    "uvicorn[standard]>=0.29.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0"]

[tool.setuptools.packages.find]
where = ["src"]
